Refactor include paths in pawgui headers to use relative paths
