+++ src/gpe_editor/editor/gpe_editor_project.h
@@ -34,9 +34,9 @@ SOFTWARE.
 #ifndef gpe_editor_project_h
 #define gpe_editor_project_h
 
-#include "../pawgui/pawgui.h"
-#include "../pawgui/pawgui_resource_dropdown.h"
-#include "../pawgui/pawgui_general_resource.h"
+#include "../pawgui/pawgui/pawgui.h"
+#include "../pawgui/resource_dropdown/pawgui_resource_dropdown.h"
+#include "../pawgui/general_resource/pawgui_general_resource.h"
 
 #include "gpe_editor_constants.h"
 #include "../editor/gpe_editor_globals.h"
