+++ src/pawgui/checkbox/pawgui_checkbox.h
@@ -35,7 +35,7 @@ Contributors:
 #ifndef PAWGUI_CHCKBOX_H
 #define PAWGUI_CHCKBOX_H
 
-#include "pawgui_base.h"
+#include "../base/pawgui_base.h"
 
 namespace pawgui
 {
