+++ src/pawgui/pawgui/pawgui.h
@@ -34,36 +34,36 @@ SOFTWARE.
 #ifndef PAWGUI_LIBRARY_H
 #define PAWGUI_LIBRARY_H
 
-#include "pawgui_base.h"
-#include "pawgui_buttons.h"
-#include "pawgui_checkbox.h"
-#include "pawgui_color_picker.h"
-#include "pawgui_content_selector.h"
-#include "pawgui_context.h"
-#include "pawgui_dropdown.h"
-#include "pawgui_file_popups.h"
-#include "pawgui_globals.h"
-#include "pawgui_labels.h"
-#include "pawgui_notifications.h"
-#include "pawgui_overlay.h"
-#include "pawgui_panel_list.h"
-#include "pawgui_parsers.h"
-#include "pawgui_popups.h"
-#include "pawgui_radiobuttons.h"
-#include "pawgui_resource_dropdown.h"
-#include "pawgui_scrollbars.h"
-#include "pawgui_search_controller.h"
-#include "pawgui_selectbox.h"
-#include "pawgui_sliders.h"
-#include "pawgui_statusbar.h"
-#include "pawgui_stree.h"
-#include "pawgui_tab_basic.h"
-#include "pawgui_tab_manager.h"
-#include "pawgui_text_editor.h"
-#include "pawgui_text_input.h"
-#include "pawgui_text_url.h"
-#include "pawgui_theme_controller.h"
-#include "pawgui_themes.h"
+#include "../base/pawgui_base.h"
+#include "../buttons/pawgui_buttons.h"
+#include "../checkbox/pawgui_checkbox.h"
+#include "../color_picker/pawgui_color_picker.h"
+#include "../context_selector/pawgui_content_selector.h"
+#include "../context/pawgui_context.h"
+#include "../dropdown/pawgui_dropdown.h"
+#include "../file_popups/pawgui_file_popups.h"
+#include "../globals/pawgui_globals.h"
+#include "../labels/pawgui_labels.h"
+#include "../notifications/pawgui_notifications.h"
+#include "../overlay/pawgui_overlay.h"
+#include "../panel_list/pawgui_panel_list.h"
+#include "../parsers/pawgui_parsers.h"
+#include "../popups/pawgui_popups.h"
+#include "../radio_buttons/pawgui_radiobuttons.h"
+#include "../resource_dropdown/pawgui_resource_dropdown.h"
+#include "../scrollbars/pawgui_scrollbars.h"
+#include "../search_controller/pawgui_search_controller.h"
+#include "../select_box/pawgui_selectbox.h"
+#include "../sliders/pawgui_sliders.h"
+#include "../statusbar/pawgui_statusbar.h"
+#include "../stree/pawgui_stree.h"
+#include "../tab_basic/pawgui_tab_basic.h"
+#include "../tab_manager/pawgui_tab_manager.h"
+#include "../text_editor/pawgui_text_editor.h"
+#include "../text_input/pawgui_text_input.h"
+#include "../text_url/pawgui_text_url.h"
+#include "../theme_controller/pawgui_theme_controller.h"
+#include "../themes/pawgui_themes.h"
 
 
 namespace pawgui
