+++ src/pawgui/scrollbars/pawgui_scrollbars.h
@@ -35,7 +35,7 @@ Contributors:
 #ifndef PAWGUI_SCROLLBARS_H
 #define PAWGUI_SCROLLBARS_H
 
-#include "pawgui_base.h"
+#include "../base/pawgui_base.h"
 
 namespace pawgui
 {
