+++ src/pawgui/stree/pawgui_stree.h
@@ -35,9 +35,9 @@ Contributors:
 #ifndef PAWGUI_STREE_H
 #define PAWGUI_STREE_H
 
-#include "pawgui_panel_list.h"
-#include "pawgui_buttons.h"
-#include "pawgui_text_input.h"
+#include "../panel_list/pawgui_panel_list.h"
+#include "../buttons/pawgui_buttons.h"
+#include "../text_input/pawgui_text_input.h"
 
 namespace pawgui
 {
